+++ src/libwatermark/io/fftproxy/window/HammingWindow.h
@@ -14,10 +14,10 @@ class HammingWindow: public WindowBase<data_type>
 
 		virtual void apply(std::vector<data_type>& v, const size_type increment) final override
 		{
-			const double f = 2.0 * M_PI / increment;
+			const double f = 2.0 * M_PI / (double)increment;
 			for(auto i = 0U; i < increment; ++i)
 			{
-				v[i] *= 0.54 - 0.46 * cos(i * f);
+				v[i] *= 0.54 - 0.46 * std::cos((double) i * f);
 			}
 		}
 };
